renaming of demos, stopping using sm_ prefix
