+++ demos/cond_var_ping.cpp
@@ -1,71 +1,81 @@
-#include "shared_memory/locked_condition_variable.hpp"
-#include "shared_memory/shared_memory.hpp"
-#include <iostream>
-#include <thread>
-
-#include <stdlib.h>
-#include <time.h>
+#include <vector>
+#include <atomic>
 #include <unistd.h>
+#include <thread>
 #include <iostream>
-#include <signal.h>
+
+#include "shared_memory/condition_variable.hpp"
+#include "shared_memory/mutex.hpp"
+#include "shared_memory/lock.hpp"
+
+#define MUTEX_ID "sm_cond_var_demo_mutex"
+#define CV_SEGMENT_ID "sm_cond_var_demo_segment"
+#define RUNNING_SEGMENT_ID "sm_running_demo_segment"
+#define RUNNING_OBJECT_ID "sm_running_demo_object"
+#define VALUE 2
 
 
-static bool RUNNING = true;
-
-void stop(int){
-  RUNNING=false;
+bool should_run(){
+  bool running;
+  shared_memory::get<bool>(RUNNING_SEGMENT_ID,
+			   RUNNING_OBJECT_ID,running);
+  return running;
 }
 
+void update_vector(){
 
-void execute(){
-
-  shared_memory::LockedConditionVariable::clean("main_memory", "cond_var");
-
-  shared_memory::LockedConditionVariable cv ("main_memory", "cond_var");
-
-  int count = 0;
-  shared_memory::set<int>("main_memory","count",count);
-
-  bool started = false;
-
-  cv.lock_scope();
   
-  while (RUNNING) {
+  std::vector<int> v(1000000);
+  shared_memory::Mutex::clean(MUTEX_ID);
+  shared_memory::clear_shared_memory(RUNNING_SEGMENT_ID);
+  shared_memory::Mutex mutex(MUTEX_ID);
+  shared_memory::ConditionVariable condition(CV_SEGMENT_ID,true);
+  shared_memory::set<bool>(RUNNING_SEGMENT_ID,
+			   RUNNING_OBJECT_ID,false);
 
-    if (!started){
-      std::cout << "waiting for pong\n";
-      started = true;
-    }
+  std::cout << "\nwaiting for pong ...\n";
+
+  bool first_iteration = true;
   
+  while(first_iteration || should_run()){
 
-    cv.wait();
+    first_iteration=false;
 
-    shared_memory::get<int>("main_memory","count",count);
-    count++;
-    shared_memory::set<int>("main_memory","count",count);
+    {
+
+      shared_memory::Lock lock(mutex);
+      condition.wait(lock);
+
+      std::cout << "PING ";
+      for(int i=0;i<v.size();i++){
+	v[i] = VALUE;
+      }
     
-    std::cout << "\t\tping " << count << "\n";
+      usleep(500);
+    
+      for(int i=0;i<v.size();i++){
+	if (v[i]!= VALUE){
+	  std::cout << "\n--- damn, the vector was not locked !\n";
+	  break;
+	}
+	if (i<10){
+	  std::cout << " " << v[i];
+	}
+      }
+      std::cout << "\n";
 
-    usleep(2000);
+    }
 
-    cv.notify_one();
+    condition.notify_one();
     
   }
 
-  cv.unlock_scope();
-
+  shared_memory::clear_shared_memory(RUNNING_SEGMENT_ID);
+  
 }
-
 
 int main(){
 
-  // cleaning and exit on ctrl+c 
-  struct sigaction cleaning;
-  cleaning.sa_handler = stop;
-  sigemptyset(&cleaning.sa_mask);
-  cleaning.sa_flags = 0;
-  sigaction(SIGINT, &cleaning, nullptr);
-
-  execute();
+  update_vector();
   
 }
