+++ demos/cond_var_pong.cpp
@@ -1,63 +1,90 @@
-#include "shared_memory/locked_condition_variable.hpp"
-#include "shared_memory/shared_memory.hpp"
-#include <iostream>
-#include <thread>
-
-#include <stdlib.h>
-#include <time.h>
+#include <vector>
+#include <atomic>
 #include <unistd.h>
+#include <thread>
 #include <iostream>
-#include <signal.h>
+
+#include "shared_memory/condition_variable.hpp"
+#include "shared_memory/mutex.hpp"
+#include "shared_memory/lock.hpp"
+
+#define MUTEX_ID "sm_cond_var_demo_mutex"
+#define CV_SEGMENT_ID "sm_cond_var_demo_segment"
+#define RUNNING_SEGMENT_ID "sm_running_demo_segment"
+#define RUNNING_OBJECT_ID "sm_running_demo_object"
+#define VALUE 1
 
 
-static bool RUNNING = true;
+void update_vector(){
 
-void stop(int){
-  RUNNING=false;
-}
-
-
-void execute(){
-
-  shared_memory::LockedConditionVariable cv ("main_memory", "cond_var",false);
-
-  cv.lock_scope();
   
-  // ping was waiting for pong to start,
-  // informing ping it can go ahead
-  cv.notify_one();
-  
-  while (RUNNING) {
-  
-    cv.wait();
+  std::vector<int> v(1000000);
+  shared_memory::Mutex::clean(MUTEX_ID);
+  shared_memory::Mutex mutex(MUTEX_ID);
+  shared_memory::ConditionVariable condition(CV_SEGMENT_ID,true);
 
-    int count;
-    shared_memory::get<int>("main_memory","count",count);
-    count++;
-    shared_memory::set<int>("main_memory","count",count);
+  // starting ping
+  std::cout << "\nstarting PING...\n";
+  condition.notify_one();
+  std::cout << "\nstarting PONG...\n";
+  
+  for(int i=0;i<10000;i++){
+
+    {
+
+      shared_memory::Lock lock(mutex);
+      condition.wait(lock);
+
+      std::cout << "PONG ";
+      for(int i=0;i<v.size();i++){
+	v[i] = VALUE;
+      }
     
-    std::cout << "\t\tpong " << count << "\n";
+      usleep(500);
+    
+      for(int i=0;i<v.size();i++){
+	if (v[i]!= VALUE){
+	  std::cout << "\n--- damn, the vector was not locked !\n";
+	  break;
+	}
+	if (i<10){
+	  std::cout << " " << v[i];
+	}
+      }
+      std::cout << "\n";
 
-    usleep(2000);
+    }
 
-    cv.notify_one();
+    condition.notify_one();
     
   }
-  
-  cv.unlock_scope();
+
+  // stopping ping
+  shared_memory::set<bool>(RUNNING_SEGMENT_ID,
+			   RUNNING_OBJECT_ID,false);
+
   
 }
-
 
 int main(){
 
-  // cleaning and exit on ctrl+c 
-  struct sigaction cleaning;
-  cleaning.sa_handler = stop;
-  sigemptyset(&cleaning.sa_mask);
-  cleaning.sa_flags = 0;
-  sigaction(SIGINT, &cleaning, nullptr);
+  bool go_ahead;
+  
+  try {
+    // ping should be started first.
+    // if it is starting, below will not fail
+    bool foo;
+    shared_memory::get<bool>(RUNNING_SEGMENT_ID,
+			     RUNNING_OBJECT_ID,foo);
+    go_ahead=true;
+  } catch(...){
+    std::cout << "\nstart sm_cond_var_ping first !\n\n";
+  }
 
-  execute();
+  if(go_ahead){
+    shared_memory::set<bool>(RUNNING_SEGMENT_ID,
+			     RUNNING_OBJECT_ID,true);
+    update_vector();
+  }
   
 }
